requirements.txt
async-timeout==5.0.1
blinker==1.9.0
boto3==1.42.70
botocore==1.42.70
cachelib==0.13.0
click==8.3.1
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.2
croniter==6.2.2
cycler==0.12.1
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.61.1
greenlet==3.3.1
holidays==0.90
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.4.9
limits==5.8.0
MarkupSafe==3.0.3
matplotlib==3.10.8
msgspec==0.20.0
narwhals==2.16.0
numpy==2.2.6
ordered-set==4.1.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.5.2
prophet==1.3.0
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
redis==7.4.0
rq==2.7.0
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.46
stanio==0.5.1
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.1.1
WTForms==3.2.1
